etl: do sort and file flush in another goroutine  (#1052)
- if provider is in-memory: do sort+flush in same goroutine
- if provider is file-based: do sort+flush in another goroutine, and
Load method will wait for unfinished goroutines and return error if one
happened inside goroutine. Also in this case do pre-palloc of new buffer
with `prevBufSize/8` size - because can't re-use prev buffer in this
case.

Reason: E4 has 8 etl collectors in same time (for
domains/history/inverted_indices) and `sort.Stable` is kind-of
bottleneck.

+++ etl/dataprovider.go
@@ -24,17 +24,20 @@ import (
 	"os"
 
 	"github.com/ledgerwatch/log/v3"
+	"golang.org/x/sync/errgroup"
 )
 
 type dataProvider interface {
 	Next(keyBuf, valBuf []byte) ([]byte, []byte, error)
-	Dispose() uint64 // Safe for repeated call, doesn't return error - means defer-friendly
+	Dispose()    // Safe for repeated call, doesn't return error - means defer-friendly
+	Wait() error // join point for async providers
 }
 
 type fileDataProvider struct {
 	file       *os.File
 	reader     io.Reader
 	byteReader io.ByteReader // Different interface to the same object as reader
+	wg         *errgroup.Group
 }
 
 // FlushToDisk - `doFsync` is true only for 'critical' collectors (which should not loose).
@@ -43,35 +46,39 @@ func FlushToDisk(logPrefix string, b Buffer, tmpdir string, doFsync bool, lvl lo
 		return nil, nil
 	}
 
-	// if we are going to create files in the system temp dir, we don't need any
-	// subfolders.
-	if tmpdir != "" {
-		if err := os.MkdirAll(tmpdir, 0755); err != nil {
-			return nil, err
+	provider := &fileDataProvider{reader: nil, wg: &errgroup.Group{}}
+	provider.wg.Go(func() error {
+		b.Sort()
+
+		// if we are going to create files in the system temp dir, we don't need any
+		// subfolders.
+		if tmpdir != "" {
+			if err := os.MkdirAll(tmpdir, 0755); err != nil {
+				return err
+			}
 		}
-	}
 
-	bufferFile, err := os.CreateTemp(tmpdir, "erigon-sortable-buf-")
-	if err != nil {
-		return nil, err
-	}
-	if doFsync {
-		defer bufferFile.Sync() //nolint:errcheck
-	}
+		bufferFile, err := os.CreateTemp(tmpdir, "erigon-sortable-buf-")
+		if err != nil {
+			return err
+		}
+		provider.file = bufferFile
 
-	w := bufio.NewWriterSize(bufferFile, BufIOSize)
-	defer w.Flush() //nolint:errcheck
+		if doFsync {
+			defer bufferFile.Sync() //nolint:errcheck
+		}
 
-	defer func() {
-		b.Reset() // run it after buf.flush and file.sync
+		w := bufio.NewWriterSize(bufferFile, BufIOSize)
+		defer w.Flush() //nolint:errcheck
+
+		if err = b.Write(w); err != nil {
+			return fmt.Errorf("error writing entries to disk: %w", err)
+		}
 		log.Log(lvl, fmt.Sprintf("[%s] Flushed buffer file", logPrefix), "name", bufferFile.Name())
-	}()
+		return nil
+	})
 
-	if err = b.Write(w); err != nil {
-		return nil, fmt.Errorf("error writing entries to disk: %w", err)
-	}
-
-	return &fileDataProvider{file: bufferFile, reader: nil}, nil
+	return provider, nil
 }
 
 func (p *fileDataProvider) Next(keyBuf, valBuf []byte) ([]byte, []byte, error) {
@@ -88,14 +95,14 @@ func (p *fileDataProvider) Next(keyBuf, valBuf []byte) ([]byte, []byte, error) {
 	return readElementFromDisk(p.reader, p.byteReader, keyBuf, valBuf)
 }
 
-func (p *fileDataProvider) Dispose() uint64 {
-	info, _ := os.Stat(p.file.Name())
-	_ = p.file.Close()
-	_ = os.Remove(p.file.Name())
-	if info == nil {
-		return 0
+func (p *fileDataProvider) Wait() error { return p.wg.Wait() }
+func (p *fileDataProvider) Dispose() {
+	if p.file != nil { //invariant: safe to call multiple time
+		p.Wait()
+		_ = p.file.Close()
+		_ = os.Remove(p.file.Name())
+		p.file = nil
 	}
-	return uint64(info.Size())
 }
 
 func (p *fileDataProvider) String() string {
@@ -161,9 +168,8 @@ func (p *memoryDataProvider) Next(keyBuf, valBuf []byte) ([]byte, []byte, error)
 	return key, value, nil
 }
 
-func (p *memoryDataProvider) Dispose() uint64 {
-	return 0 /* doesn't take space on disk */
-}
+func (p *memoryDataProvider) Wait() error { return nil }
+func (p *memoryDataProvider) Dispose()    {}
 
 func (p *memoryDataProvider) String() string {
 	return fmt.Sprintf("%T(buffer.Len: %d)", p, p.buffer.Len())
